requirements.txt
imgviz==1.7.2
labelme==5.2.1
numpy==1.24.3
Pillow==9.5.0
scipy==1.10.1
